==> 4.분류/1.Bus or Raid.ipynb <==
import pandas as pd

# trade 데이터로부터 이상값들을 추적한다.

tr = pd.read_csv('../preprocess/train_trade_pre.csv')
te = pd.read_csv('../preprocess/test_trade_pre.csv')
to = pd.concat([tr, te]).drop_duplicates().sort_values(by=list(tr.columns)).reset_index(drop=True)
to.columns

# auto 계정들과, 버스기사, 골드 구매자를 구분하기 위한 과정이다.
# source가 target에게 여러가지 재화를 지급한 목록을 구한다. 

dup = to[to.duplicated(['trade_week','trade_day','trade_time','source_acc_id','target_acc_id'])]

sou = pd.DataFrame(dup['source_acc_id'].unique())
tar = pd.DataFrame(dup['target_acc_id'].unique())
sou.rename(columns={0:'acc_id'}, inplace=True)
tar.rename(columns={0:'acc_id'}, inplace=True)
sou = sou.sort_values(by='acc_id').reset_index(drop=True)
tar = tar.sort_values(by='acc_id').reset_index(drop=True)

print('Number of duplicated Source : {}\nNumber of duplicated Target : {}'.format(len(sou), len(tar)))

# Auto 분류는 sou, tar 데이터를 이용하여 2번 스크립트에서 추려내며,
# 현재 1번 스크립트에서는 버스기사나 레이드 활동자를 추려낸다.

tar.to_csv('auto_target_dup.csv', index=False)
sou = pd.DataFrame(set(sou['acc_id']) - set(tar['acc_id'])).rename(columns={0:'acc_id'}).sort_values(by='acc_id')
sou.to_csv('auto_source_dup.csv', index=False)

# 버스기사 또는 레이드 이익잉여금 분배는 다음과 같이 이뤄진다.
* 1명의 Source가 money를 같은 금액으로 다수의 Target과 나눠가진다.
* 때문에 duplicated 조건에 source_acc_id와 item_amount가 반드시 포함되어야 한다.

bus_or_raid = to[to.duplicated(['trade_week', 'source_acc_id', 'item_type', 'item_amount'])]
bus_or_raid.head()

money_trader = bus_or_raid[bus_or_raid['item_type']=='money']
driver_receiver = sorted(set(money_trader['source_acc_id']) | set(money_trader['target_acc_id']))
b_or_r = pd.DataFrame(driver_receiver).rename(columns={0: 'acc_id'})

driver = pd.merge(to, b_or_r.rename(columns={'acc_id': 'source_acc_id'}), how='inner').sort_values(by=list(to.columns))

### 1000 금 이상 오간 데이터만 확인한다.

temp = driver[(driver['item_type']=='money') & (driver['item_amount']>=1000)].reset_index(drop=True)
temp.head()

sou_acc = temp['source_acc_id'].unique()
tar_acc = temp['target_acc_id'].unique()
print('Number of Large Source : {}\nNumber of Large Target : {}'.format(len(sou_acc), len(tar_acc)))

print('Large Trader : {}'.format(len(set(sou_acc) | set(tar_acc))))

### 게임에서 레이드 분배금을 나눠가지거나 큰 금액을 거래하는 이들은 32,872개.
### 이 중 이후 판별될 auto 관리자 계정을 제거하면 게임을 열심히 즐기고 있는 유저로 생각 할 수 있다.

hard_list = pd.DataFrame(sorted(set(sou_acc) | set(tar_acc))).rename(columns={0:'acc_id'})
hard_list.head()

hard_list.to_csv('hard_list.csv', index=False)
