==> 4.분류/2.Auto.ipynb <==
import pandas as pd

# auto 계정을 추려내기 위해서 다음의 데이터를 사용한다.
* 중복거래 데이터
* 활동데이터

tr = pd.read_csv('../3.통합/4.주당평균_파생변수/train_create_mean.csv')
te = pd.read_csv('../3.통합/4.주당평균_파생변수/test_create_mean.csv')
to = pd.concat([tr, te], sort=False).reset_index(drop=True)
source = pd.read_csv('auto_source_dup.csv')
target = pd.read_csv('auto_target_dup.csv')
hard = pd.read_csv('hard_list.csv')

to_source = pd.merge(to, source, on='acc_id', how='inner')
to_target = pd.merge(to, target, on='acc_id', how='inner')
to_hard = pd.merge(to, hard, on='acc_id', how='inner')

print(list(to.columns))

# and 조건
sou_res = to_source[
    # 개미 계정은 레이드 참가가 불가능하다.
    (to_source['avg_cnt_enter_raid']==0)
    & (to_source['avg_cnt_enter_raid_light']==0)
    & (to_source['avg_cnt_enter_bam']==0)
    # 개미 계정은 캐쉬 충전을 하지 않는다.
    & (to_source['avg_payment_amount']==0)
    # 개미 계정은 돈과 잡화만 관리자에게 공급한다.
    & (to_source['avg_give_money_count']>0)
    & (to_source['avg_give_grocery_count']>0)
    & (to_source['avg_give_weapon_count']==0)
    & (to_source['avg_give_accessory_count']==0)
    & (to_source['avg_give_gem_count']==0)
    & (to_source['avg_give_costume_count']==0)
    # 개미 계정은 전투를 통해 재화를 번다.
    & (to_source['avg_game_combat_time']>=30)
    # 개미 계정은 길드가 없음
    & (to_source['guild_count']==0)
    # 동일 관리자 계정에 반복하여 재화를 공급한다.
    & (to_source['dup_give_count']>0)
    & (to_source['avg_give_money_sum']>to_source['avg_rec_money_sum'])
    & (to_source['avg_give_grocery_sum']>to_source['avg_rec_grocery_sum'])
]

# or 조건
sou_res = sou_res[
    (sou_res['avg_cnt_enter_inzone_solo']>0)
    | (sou_res['avg_cnt_enter_inzone_light']>0)
    | (sou_res['avg_cnt_enter_inzone_skilled']>0)
    | (sou_res['avg_cnt_enter_inzone_normal']>0)
].reset_index(drop=True)

# auto 기준과 hard 기준의 중복 계정
# 해당 계정들은 hard 유저가 아닌 auto 계정으로 판별됨.
auto_hard = pd.merge(sou_res, hard, on='acc_id', how='inner')
auto = auto_hard[auto_hard['avg_trade_count']>5]
hard = pd.DataFrame(set(hard['acc_id']) - set(auto['acc_id'])).rename(columns={0: 'acc_id'})
auto = pd.DataFrame(set(sou_res['acc_id']) - set(hard['acc_id'])).rename(columns={0: 'acc_id'})

auto['ant_id'] = 1
auto['master_id'] = 0
auto.head()

tar_res = to_target[
    (to_target['avg_cnt_enter_raid']==0)
    & (to_target['avg_cnt_enter_raid_light']==0)
    & (to_target['avg_payment_amount']==0)
    & (to_target['avg_rec_count']>10)
    & (to_target['avg_rec_money_sum']>1000)
    & (to_target['avg_rec_grocery_count']>0)
    & (to_target['dup_rec_acc_count']>1)
]

master = pd.DataFrame(tar_res['acc_id']).reset_index(drop=True).rename(columns={0: 'acc_id'})
hard = pd.DataFrame(set(hard['acc_id']) - set(master['acc_id'])).rename(columns={0: 'acc_id'})
master['ant_id'] = 0
master['master_id'] = 1
master.head()

auto_list = pd.concat([auto, master]).sort_values(by='acc_id').reset_index(drop=True)
print(auto_list.groupby(['ant_id', 'master_id']).count())
auto_list['auto_id']=1

auto_list.to_csv('auto_list.csv', index=False)
hard.to_csv('hard_list_remove_auto.csv', index=False)
