==> 4.분류/3.Dummy_8610.ipynb <==
import pandas as pd

# Dummy 계정 분류
Dummy 계정은 튜토리얼만 클리어 한 계정이다.  
대나무 마을에서 은광일 지부장을 잡는 퀘스트까지가 튜토리얼이며 이를 완료하면 거래가 가능하다.  
따라서 이들은 거래만을 위해서 만들어진 Dummy 계정으로 생각할 수 있다.

**이 때 얻는 퀘스트 경험치 총량은 8610이다.**

tr = pd.read_csv('../3.통합/4.주당평균_파생변수/train_create_mean.csv')
te = pd.read_csv('../3.통합/4.주당평균_파생변수/test_create_mean.csv')
to = pd.concat([tr, te], sort=False)

to_8610 = to[to['activated_wk']*to['avg_quest_exp']==8610].reset_index(drop=True)
to_8610['8610_id'] = 1
to_8610 = to_8610[['acc_id', '8610_id']]

to_8610.to_csv('8610_list.csv', index=False)
