==> 4.분류/4.Hard_list.ipynb <==
import pandas as pd

tr = pd.read_csv('../3.통합/4.주당평균_파생변수/train_create_mean.csv')
te = pd.read_csv('../3.통합/4.주당평균_파생변수/test_create_mean.csv')
to = pd.concat([tr, te], sort=False)

hard = pd.read_csv('hard_list_remove_auto.csv')

to_hard = pd.merge(to, hard, on='acc_id', how='inner')
to_hard['hard_id'] = 1
to_hard = to_hard[['acc_id', 'hard_id']]
len(to_hard)

to_hard.to_csv('hard_list_fin.csv', index=False)
