==> 4.분류/5.Jump_list.ipynb <==
import pandas as pd

# Jumping Character
고속성장이용권을 사용하여 캐릭터를 생성했을 때 튜토리얼을 완료하면 홍문 퀘스트 경험치 4,500,000을 얻는다.  
이를 따로 분류한다.

tr = pd.read_csv('../3.통합/4.주당평균_파생변수/train_create_mean.csv')
te = pd.read_csv('../3.통합/4.주당평균_파생변수/test_create_mean.csv')
to = pd.concat([tr, te], sort=False)

to_jump = to[to['activated_wk']*to['avg_quest_hongmun']==4.5e6].reset_index(drop=True)
to_jump['jump_id'] = 1
to_jump = to_jump[['acc_id', 'jump_id']]

to_jump.to_csv('jump_list.csv', index=False)
